requirements.txt
# Python Dependencies for Logo Detection & Recognition System

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLOv8
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Handling
pandas>=2.0.0
PyYAML>=6.0

# Progress Bars
tqdm>=4.65.0

# COCO Tools (for dataset handling)
pycocotools>=2.0.6

# TensorBoard (for training monitoring)
tensorboard>=2.13.0

# Jupyter (for notebook training)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0

# Optional: ONNX export
# onnx>=1.14.0
# onnxruntime>=1.15.0
